+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## 3.0.0 - 2020-02-14
+- Rename package name.
+
 ## 2.0.3 - 2020-02-14
 - Improved Exception.
 
@@ -43,4 +46,4 @@
 - Improved design.
 
 ## 1.0.0 - 2019-05-02
-- Initial release of the HttpX Tea Version 1.0.0 on Packagist See <https://github.com/aliyun/tea-php> for more information.
+- Initial release of the AlibabaCloud Tea Version 1.0.0 on Packagist See <https://github.com/aliyun/tea-php> for more information.